backend/requirements.txt
autopep8==1.6.0
boto3==1.26.6
celery==5.1.2
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0
dj-rest-auth==2.2.4
django-allauth==0.50.0
django-anymail==8.5
django-cors-headers==3.10.1
django-environ==0.8.1
django-extensions==3.1.5
django-filter==21.1
django-rest-auth==0.9.5
django-rest-swagger==2.2.0
django-storages==1.12.3
django-nested-inline==0.4.6
Django==3.2.12
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-yasg==1.20.0
gevent==21.12.0
greenlet==1.1.2
gunicorn==20.1.0
pillow==10.0.0
psycopg2-binary==2.8.6
pyfcm==1.5.4
python-docx==0.8.11
redis==4.5.3
Werkzeug==2.0.3
zope.interface==5.4.0
zope.event==4.5.0
channels-auth-token-middlewares
